- forgot one Py_ssize_t def

+++ src/pycurl.c
@@ -1164,7 +1164,7 @@ read_callback(char *ptr, size_t size, size_t nmemb, void *stream)
     /* handle result */
     if (PyString_Check(result)) {
         char *buf = NULL;
-        int obj_size = -1;
+        Py_ssize_t obj_size = -1;
         Py_ssize_t r;
         r = PyString_AsStringAndSize(result, &buf, &obj_size);
         if (r != 0 || obj_size < 0 || obj_size > total_size) {
